Rewrote Map Room Generation
Changed how the rooms are generated. Still have to make some adjustments, like for the small rooms to make it so that they are split with a single wall instead of a corridor between them. Also in Map.cs have to make a few adjustments to make sure that all the rooms have doors and can be accessed - some kind of fill function like the last one.