Minor Improvements
Restructured most of the folder hierarchy.

Fixed the bug where projectiles would fly through doors.

Added the MapDecorator class, with the role to fill the map with objects. Not implemented yet.

Added a floor interactible prefab with animations. Will be either a repair station or energy charging station. But first I have to create a game manager which is scene-independent, where I will handle a lot of common interactions.