Improved player and shooting, added energy system
Player script (with movement and health etc) is now working with the modifiers (for exmaple I can set the movement modifier and the player moves slower). Also made functional the health modifying system.

Added an energy system to player: energy is lost every second, more energy is lost while moving, and for now, energy is lost while shooting(since it's a laser weapon for now). All customizable from other scripts with modifiers.

Updated the UI to have better looking health display, and added a similar display for energy.

Finally, added an energy charging prefab, which is designed to be placed on the floor, and the player can run on top of it. It charges the player's energy at a customizable rate, and has animations for idle, charging and depleted.